PlayerFSM movement bug fix
Since we use OnAirMovementState as a child state of PlayerJumpState and PlayerFallingState, we need to synchronize the position before initiating movement. Otherwise, the OnAirMovementState will move the character, and the PlayerJumpState and PlayerFallingState will not have the updated position. This was causing problems on low FPS when the character was near Ground Layer. Couldn't exprience this on target frame rate (60)

Note: We are currently have a bug when the player is near the Ground Layer. Sometimes, the GroundedState resets the player's position to the top of a long wall. While this behavior may be intended when entering the GroundedState, the FSM should never transition to this state if there is no ground underneath the character.