SceneLoading visual tweaks, used 2 cameras
Now we use multiple cameras just for displaying scene loading, this is not what I want but setting the PersistantScene camera to Depth only causes it to not clear depth buffer. We need another camera that clears the depth in order to display correctly. It works as expected now but didn't like this approach.